+++ R/AzureRMR.R
@@ -5,6 +5,11 @@ NULL
 {
     azure_api_version="2018-05-01"
     options(azure_api_version=azure_api_version)
+
+    config_dir <- config_dir()
+    if(!dir.exists(config_dir))
+        dir.create(config_dir, recursive=TRUE)
+
     invisible(NULL)
 }