+++ R/az_subscription.R
@@ -26,8 +26,7 @@
 #' \dontrun{
 #'
 #' # recommended way to retrieve a subscription object
-#' sub <- az_rm$
-#'     new(tenant="myaadtenant.onmicrosoft.com", app="app_id", password="password")$
+#' sub <- get_azure_login("myaadtenant")$
 #'     get_subscription("subscription_id")
 #'
 #' # retrieve list of resource group objects under this subscription